+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 
+ * Fix `Parser` incorrectly translating non-documentation comments as part of documentation comments ([issue #475](https://github.com/bytedeco/javacpp/issues/475))
+ * Set `Pointer.maxPhysicalBytes` to `4 * Runtime.maxMemory()` by default as workaround for memory-mapped files, ZGC, etc ([issue #468](https://github.com/bytedeco/javacpp/issues/468))
  * Ensure `synchronized` code in `Pointer` gets skipped with "org.bytedeco.javacpp.nopointergc" ([issue tensorflow/java#313](https://github.com/tensorflow/java/issues/313))
  * Add `protected Pointer.offsetAddress()` and use it for `getPointer()` instead of `position()`
  * Fix potential infinite loop in `Parser` when processing `class`, `struct`, or `union` declarations